validaciones producto y modificar producto

+++ src/main/resources/static/js/productos.js
@@ -1,13 +1,13 @@
-        function getProductos() {
-            axios.get("/producto/listarProductos")
-            .then(function(response) {
-                const productos = response.data;
-                const tbody = document.querySelector("tbody");
-                let htmlContent = '';
-                productos.forEach(producto => {
-                    // Comprobar que producto, marca y categoría existen
-                    if (producto && producto.marca && producto.categoria) {
-                        htmlContent += `
+function getProductos() {
+    axios.get("/producto/listarProductos")
+        .then(function (response) {
+            const productos = response.data;
+            const tbody = document.querySelector("tbody");
+            let htmlContent = '';
+            productos.forEach(producto => {
+                // Comprobar que producto, marca y categoría existen
+                if (producto && producto.marca && producto.categoria) {
+                    htmlContent += `
                             <tr>
                                 <td>${producto.nombre}</td>
                                 <td>${producto.descripcion}</td>
@@ -20,38 +20,67 @@
                                     ${producto.categoria.nombre}
                                 </td>
                                 <td><button class="btn-modificar" onclick="location.href='/modificarProducto/${producto.id}'">Modificar</button></td>
-                                <td><button class="btn-eliminar" onclick="eliminarProducto(${producto.id})">Eliminar</button></td>
+                                <td><button class="btn-eliminar" onclick="eliminarProducto(${producto.id}, '${producto.nombre}')">Eliminar</button></td>
                             </tr>
                         `;
-                    } else {
-                        console.warn('Producto o propiedades faltantes:', producto);
-                    }
+                } else {
+                    console.warn('Producto o propiedades faltantes:', producto);
+                }
+            });
+
+            tbody.innerHTML = htmlContent;
+        })
+        .catch((err) => console.error(err));
+}
+
+function eliminarProducto(id, nombre) {
+
+    Swal.fire({
+        title: "Está seguro de eliminar este producto?",
+        text: "Se eliminará: " + nombre,
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Sí, eliminar",
+        cancelButtonText: "Cancelar"
+    }).then((result) => {
+        if (result.isConfirmed) {
+
+            axios.delete(`/producto/eliminarProducto/${id}`)
+                .then(response => {
+
+                    Swal.fire({
+                        position: "top-center",
+                        icon: "success",
+                        title: "Producto eliminado correctamente!",
+                        showConfirmButton: false,
+                        timer: 1500
+                    });
+
+                    setTimeout(function () {
+                        window.location.href = '/registrarProducto';
+                    }, 1500);
+                })
+                .catch(error => {
+                    console.error('Error:', error);
                 });
 
-                tbody.innerHTML = htmlContent;
-            })
-            .catch((err) => console.error(err));
+        } else {
+            return;
         }
-        
-        function eliminarProducto(id) {
-            axios.delete(`/producto/eliminarProducto/${id}`)
-            .then(response => {
-                console.log('Éxito:', response.data);
-                alert('Producto Eliminado exitosamente!');
-                window.location.href='/registrarProducto';
-            })
-            .catch(error => {
-                console.error('Error:', error);
-            });
-        }
-        function getProductoById(id) {
-            axios.get(`/producto/obtenerProductoPorId/${id}`)
-            .then(function(response) {
-                const producto = response.data;
-                const tbody = document.querySelector("tbody");  
-                let htmlContent = '';
-                if (producto) {
-                    htmlContent += `
+    });
+
+
+}
+function getProductoById(id) {
+    axios.get(`/producto/obtenerProductoPorId/${id}`)
+        .then(function (response) {
+            const producto = response.data;
+            const tbody = document.querySelector("tbody");
+            let htmlContent = '';
+            if (producto) {
+                htmlContent += `
                         <tr>
                             <td>${producto.nombre}</td>
                             <td>${producto.descripcion}</td>
@@ -68,26 +97,26 @@
                             <td><button onclick="eliminarProducto(${producto.id})">Eliminar</button></td>
                         </tr>
                     `;
-                }
-                tbody.innerHTML = htmlContent;
-            })
-            .catch((err) => console.error(err));
-        }
-        
+            }
+            tbody.innerHTML = htmlContent;
+        })
+        .catch((err) => console.error(err));
+}
 
-        document.addEventListener("DOMContentLoaded", function() {
-            const filtroInput = document.getElementById('filtro');
-            
-            // Cargar usuarios al inicio
+
+document.addEventListener("DOMContentLoaded", function () {
+    const filtroInput = document.getElementById('filtro');
+
+    // Cargar usuarios al inicio
+    getProductos();
+
+    // Evento de entrada para el campo de filtro
+    filtroInput.addEventListener('input', function () {
+        const filtroValue = filtroInput.value;
+        if (filtroValue.trim() !== "") {
+            getProductoById(filtroValue);
+        } else {
             getProductos();
-            
-            // Evento de entrada para el campo de filtro
-            filtroInput.addEventListener('input', function() {
-                const filtroValue = filtroInput.value;
-                if (filtroValue.trim() !== "") {
-                    getProductoById(filtroValue);
-                } else {
-                    getProductos();
-                }
-            });
-        });
+        }
+    });
+});
